2019 - 11.1 + 11.2 C++
I originally solved this on the 11th, but as I had broken my IntCode VM
I had to borrow someone's VM to solve the problem quickly.

In this case I used the day09.py from
https://github.com/benediktwerner/AdventOfCode/tree/74b00d79a5da05eb2919fd2000ce3bb8b46f33e5

This is a quite literal translation of my python code, now rebased on my
day09/day13 vm.